python-client/requirements.txt
dbus-python==1.2.18
distro==1.7.0
protobuf==3.20.0
PyGObject==3.42.1
pyzmq==25.1.2
ssh-import-id==5.11